+++ src/main/java/com/example/aniwhere/service/recommend/ClusteringService.java
@@ -2,7 +2,7 @@ package com.example.aniwhere.service.recommend;
 
 import com.example.aniwhere.domain.anime.Anime;
 import com.example.aniwhere.domain.division.Division;
-import com.example.aniwhere.domain.division.DivisionRepository;
+import com.example.aniwhere.repository.DivisionRepository;
 import com.example.aniwhere.domain.user.User;
 import com.example.aniwhere.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
